+++ samples/SessionSample/project.json
@@ -4,7 +4,7 @@
     "dependencies": {
         "Microsoft.AspNet.Http.Extensions": "1.0.0-*",
         "Microsoft.AspNet.Server.IIS": "1.0.0-*",
-        "Microsoft.AspNet.Cache.Session": "1.0.0-*",
+        "Microsoft.AspNet.Session": "1.0.0-*",
         "Microsoft.Framework.Cache.Redis": "1.0.0-*"
     },
     "frameworks" : {
