--- src/Microsoft.AspNet.Cache.Session/ISessionStore.cs
@@ -1,15 +1,0 @@
-﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using Microsoft.AspNet.HttpFeature;
-
-namespace Microsoft.AspNet.Cache.Session
-{
-    public interface ISessionStore
-    {
-        bool IsAvailable { get; }
-        void Connect();
-        ISession Create(string sessionId, TimeSpan idleTimeout, Func<bool> tryEstablishSession);
-    }
-}
