+++ src/Microsoft.AspNet.Session/CachingServicesExtensions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.AspNet.Cache.Session;
+using Microsoft.AspNet.Session;
 
 namespace Microsoft.Framework.DependencyInjection
 {
