+++ src/Microsoft.AspNet.Session/DistributedSessionStore.cs
@@ -5,7 +5,7 @@ using System;
 using Microsoft.AspNet.HttpFeature;
 using Microsoft.Framework.Cache.Distributed;
 
-namespace Microsoft.AspNet.Cache.Session
+namespace Microsoft.AspNet.Session
 {
     public class DistributedSessionStore : ISessionStore
     {
