+++ src/Microsoft.AspNet.Session/Microsoft.AspNet.Session.kproj
@@ -8,22 +8,20 @@
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
   </PropertyGroup>
-
   <Import Project="$(VSToolsPath)\AspNet\Microsoft.Web.AspNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>71802736-f640-4733-9671-02d267edd76a</ProjectGuid>
     <OutputType>Library</OutputType>
-    <RootNamespace>Microsoft.AspNet.Cache.Session</RootNamespace>
+    <RootNamespace>Microsoft.AspNet.Session</RootNamespace>
   </PropertyGroup>
-
   <PropertyGroup Condition="$(OutputType) == 'Console'">
     <DebuggerFlavor>ConsoleDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup Condition="$(OutputType) == 'Web'">
     <DebuggerFlavor>WebDebugger</DebuggerFlavor>
   </PropertyGroup>
-
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'" Label="Configuration">
+    <AssemblyName>Microsoft.AspNet.Session</AssemblyName>
   </PropertyGroup>
   <PropertyGroup>
     <SchemaVersion>2.0</SchemaVersion>
