+++ src/Microsoft.AspNet.Session/SessionDefaults.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.AspNet.Cache.Session
+namespace Microsoft.AspNet.Session
 {
     public static class SessionDefaults
     {
