--- test/Microsoft.AspNet.Cache.Session.Tests/project.json
@@ -1,18 +1,0 @@
-{
-    "dependencies": {
-        "Microsoft.AspNet.Cache.Session": "1.0.0-*",
-        "Microsoft.AspNet.Http.Extensions": "1.0.0-*",
-        "Microsoft.AspNet.TestHost": "1.0.0-*",
-        "Xunit.KRunner": "1.0.0-*"
-    },
-    "commands": {
-        "test": "Xunit.KRunner"
-    },
-    "frameworks": {
-        "aspnet50": {
-            "dependencies": {
-
-            }
-        }
-    }
-}
