+++ test/Microsoft.AspNet.Session.Tests/Microsoft.AspNet.Session.Tests.kproj
@@ -8,22 +8,20 @@
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
   </PropertyGroup>
-
   <Import Project="$(VSToolsPath)\AspNet\Microsoft.Web.AspNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>8c131a0a-bc1a-4cf3-8b77-8813fbfe5639</ProjectGuid>
     <OutputType>Library</OutputType>
-    <RootNamespace>Microsoft.AspNet.Cache.Session</RootNamespace>
+    <RootNamespace>Microsoft.AspNet.Session</RootNamespace>
   </PropertyGroup>
-
   <PropertyGroup Condition="$(OutputType) == 'Console'">
     <DebuggerFlavor>ConsoleDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup Condition="$(OutputType) == 'Web'">
     <DebuggerFlavor>WebDebugger</DebuggerFlavor>
   </PropertyGroup>
-
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'" Label="Configuration">
+    <AssemblyName>Microsoft.AspNet.Session.Tests</AssemblyName>
   </PropertyGroup>
   <PropertyGroup>
     <SchemaVersion>2.0</SchemaVersion>
