+++ test/Microsoft.AspNet.Session.Tests/project.json
@@ -1,0 +1,17 @@
+{
+  "dependencies": {
+    "Microsoft.AspNet.Session": "1.0.0-*",
+    "Microsoft.AspNet.Http.Extensions": "1.0.0-*",
+    "Microsoft.AspNet.TestHost": "1.0.0-*",
+    "Xunit.KRunner": "1.0.0-*"
+  },
+  "commands": {
+    "test": "Xunit.KRunner"
+  },
+  "frameworks": {
+    "aspnet50": {
+      "dependencies": {}
+    }
+  },
+  "webroot": null
+}
